requirements.txt
tensorflow
streamlit
numpy
pillow
matplotlib
scikit-learn
scipy